pyproject.toml
[project]
name = "kose-kata"
version = "0.1.0"
description = "Powerful note Discord bot"
authors = [{ name = "Mazzie", email = "mazziechai@gmail.com" }]
dependencies = [
    "hikari[speedups]>=2.0.0.dev121",
    "sqlalchemy[asyncio]>=2.0.30",
    "uvloop>=0.17.0",
    "hikari-miru>=3.2.0",
    "asyncpg>=0.28.0",
    "alembic>=1.12.0",
    "aiosqlite>=0.20.0",
    "hikari-arc>=1.3.0",
    "python-dotenv>=1.0.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "MPL-2.0" }

[tool.pdm.dev-dependencies]
dev = ["pre-commit>=3.4.0", "black>=23.9.1", "ruff>=0.4.4"]

[tool.pdm]
distribution = false

[tool.isort]
profile = "black"
multi_line_output = 3
